fix: binding required

+++ models/auth_model.go
@@ -34,8 +34,8 @@ type User struct {
 	CreatedAt      time.Time   `bson:"createdAt"`
 	UpdatedAt      time.Time   `bson:"updatedAt"`
 	AvatarUrl      string      `bson:"avatarUrl"`
-	Lat            float64     `bson:"lat"`
-	Lng            float64     `bson:"lng"`
+	Lat            float32     `bson:"lat"`
+	Lng            float32     `bson:"lng"`
 	TrainerInfo    TrainerInfo `bson:"trainerInfo,omitempty"`
 }
 
@@ -77,7 +77,7 @@ func (e *UserAlreadyExist) Error() string {
 	return "error: user already existed"
 }
 
-func CreateUser(username string, password string, userType string, firstName string, lastName string, birthDate string, citizenID string, gender string, phoneNumber string, address string, avatarUrl string, lat float64, lng float64) (user User, err error) {
+func CreateUser(username string, password string, userType string, firstName string, lastName string, birthDate string, citizenID string, gender string, phoneNumber string, address string, avatarUrl string, lat float32, lng float32) (user User, err error) {
 	ctx, cancel := context.WithTimeout(context.Background(), 10*time.Second)
 	defer cancel()
